# cmd/hzpaste-api/internal/docs/swagger.yaml
definitions:
  paste.Paste:
    properties:
      body:
        type: string
      created:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - body
    - name
    type: object
info:
  contact:
    email: ep4sh2k@gm[a]il.com
    name: API Support
    url: http://ep4sh.cc
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /pastes/{id}:
    get:
      consumes:
      - application/json
      description: get paste by ID
      operationId: get-string-by-int
      parameters:
      - description: Paste ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: paste
          schema:
            $ref: '#/definitions/paste.Paste'
      summary: finds a single paste identified by an ID in the request URL.
swagger: "2.0"